api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.55.0
pdfplumber>=0.11.0
